games/snake/pyproject.toml
[project]
name = "Snake"
version = "0.1.0"
description = "Eat apples. Watch out for yourself."
requires-python = ">=3.10,<4.0"
dependencies = [
    "retro-games==1.2.0",
]

[project.scripts]
play = "snake:main"

[tool.retro]
author = "Chris"
description = "Eat apples. Watch out for yourself."
instructions = "Use the arrow keys to steer the snake."
result_file = "result.json"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
